+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
+## 1.1.0
+
+* Added `FenceChunk` class.
+	* Also added this class to the `defaultChunkRegistry` for 0x3000007 chunks.
+* Added `Fence2Chunk` class.
+	* Also added this class to the `defaultChunkRegistry` for 0x3000000 chunks.
+
 ## 1.0.0
 Initial release.